Typescript camera console (#78412)
## About The Pull Request
Refactors the camera console into typescript, including the secureye,
removing the need for its custom css file


![KaBYZCTqG1](https://github.com/tgstation/tgstation/assets/42397676/4de52792-943c-4b2d-94e5-8ff10b678a91)

![Screenshot 2023-09-17
120550](https://github.com/tgstation/tgstation/assets/42397676/a5370678-ffc0-44b4-94b7-f3de95d8286e)
## Why It's Good For The Game
Spruces up the UI ever so slightly and possibly fixes an issue #78009

Less code is better code
## Changelog
:cl:
refactor: Refactors the camera console UI.
/:cl:
