--- tgui/packages/tgui/styles/interfaces/CameraConsole.scss
@@ -1,53 +1,0 @@
-@use '../base.scss';
-
-$background-color: rgba(0, 0, 0, 0.33) !default;
-
-.CameraConsole__left {
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  width: base.em(220px);
-}
-
-.CameraConsole__right {
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  left: base.em(220px);
-  right: 0;
-  background-color: $background-color;
-}
-
-.CameraConsole__toolbar {
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  height: 2em;
-  line-height: 2em;
-  margin: 0.25em 1em 0;
-}
-
-.CameraConsole__toolbarRight {
-  position: absolute;
-  top: 0;
-  right: 0;
-  height: 2em;
-  line-height: 2em;
-  margin: 0.33em 0.5em 0;
-}
-
-.CameraConsole__map {
-  position: absolute;
-  top: base.em(26px);
-  bottom: 0;
-  left: 0;
-  right: 0;
-  margin: 0.5em;
-  text-align: center;
-
-  .NoticeBox {
-    margin-top: calc(50% - 2em);
-  }
-}
